.github/requirements.txt
lxml==4.6.1
translate-toolkit==3.2.0
